fix: minor refactoring
