+++ packages/amplify-schema-validator/src/__tests__/schemas/invalid-field-not-in-related-model.graphql
