+++ packages/amplify-schema-validator/src/__tests__/schemas/valid-field-in-related-model.graphql
