+++ packages/amplify-schema-validator/src/__tests__/validates-fields-match-in-related-model.test.ts
@@ -3,13 +3,13 @@ import { readSchema } from './helpers/readSchema';
 
 describe('Validate Schema', () => {
   it('fails validation when field is not in the parent model', () => {
-    const schema = readSchema('invalid-field-not-in-parent-model.graphql');
+    const schema = readSchema('invalid-field-not-in-related-model.graphql');
     const errorRegex = 'email is not a field in Test1';
     expect(() => validateSchema(schema)).toThrow(errorRegex);
   });
 
   it('passes validation when field is in the parent model', () => {
-    const schema = readSchema('valid-field-in-parent-model.graphql');
+    const schema = readSchema('valid-field-in-related-model.graphql');
     expect(() => validateSchema(schema)).not.toThrow();
   });
 });
