+++ packages/amplify-schema-validator/src/index.ts
@@ -5,7 +5,7 @@ import { parse } from 'graphql';
 import { validateIndexScalarTypes } from './validators/index-scalar-types';
 import { validateRequireBelongsToRelation } from './validators/relation-required-with-belongs-to';
 import { validateManyToManyTwoLocations } from './validators/two-many-to-many-locations';
-import { validateFieldInParentModel } from './validators/fields-passed-are-in-parent-model';
+import { validateFieldsMatchInRelatedModel } from './validators/fields-match-in-related-model';
 import { ValidationError } from './exceptions/validation-error';
 import { validateFieldIsDefinedOnce } from './validators/field-is-defined-once';
 
@@ -13,7 +13,7 @@ const allValidators = [
   validateIndexScalarTypes,
   validateRequireBelongsToRelation,
   validateManyToManyTwoLocations,
-  validateFieldInParentModel,
+  validateFieldsMatchInRelatedModel,
   validateFieldIsDefinedOnce,
 ];
